requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncio==3.4.3
attrs==24.2.0
aws-lambda-powertools==1.25.5
aws-xray-sdk==2.14.0
boto3==1.35.57
botocore==1.35.57
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.4
fastjsonschema==2.20.0
frozenlist==1.5.0
fyers_apiv3==3.1.4
h11==0.14.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.4
jmespath==0.10.0
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.1.3
pandas==2.2.3
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
requests==2.31.0
s3transfer==0.10.3
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websocket-client==1.6.1
websockets==14.0
wrapt==1.16.0
yarl==1.17.1
